refactor(portal-listing-validation): collapse ValidatePortalListingDomainService to concrete core @DomainService
