+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/test/java/io/gravitee/apim/core/portal_listing/use_case/ValidatePortalListingUseCaseTest.java
@@ -17,10 +17,10 @@ package io.gravitee.apim.core.portal_listing.use_case;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import inmemory.ValidatePortalListingDomainServiceInMemory;
 import io.gravitee.apim.core.audit.model.AuditActor;
 import io.gravitee.apim.core.audit.model.AuditInfo;
 import io.gravitee.apim.core.portal.model.PortalId;
+import io.gravitee.apim.core.portal_listing.domain_service.ValidatePortalListingDomainService;
 import io.gravitee.apim.core.portal_listing.model.PortalListingApiEntry;
 import io.gravitee.apim.core.portal_listing.model.PortalListingId;
 import io.gravitee.rest.api.service.common.HRIDToUUID;
@@ -45,7 +45,7 @@ class ValidatePortalListingUseCaseTest {
         HRIDToUUID.portalListing().context(AUDIT_INFO).portal(PORTAL_HRID).hrid(LISTING_HRID).id()
     );
 
-    private final ValidatePortalListingDomainServiceInMemory validator = new ValidatePortalListingDomainServiceInMemory();
+    private final ValidatePortalListingDomainService validator = new ValidatePortalListingDomainService();
     private ValidatePortalListingUseCase useCase;
 
     @BeforeEach
